Refactor
Complete.
