+++ levels.py
@@ -1,0 +1,31 @@
+levels = {
+1 : [
+    [0,4], [0,7], [2,2], [3,0], [3,6], [4,0], [4,3], [4,2], [4,4], [4,5], [4,6], [4,9], [5,4], [8,7], [9,5],[9,6], [9,7]
+],
+2 : [
+    [0,4], [0,0], [0,9], [2,3], [2,4], [2,6], [3,6], [3,0], [4,0], [4,4], [4,6], [4,9], [5,4], [6,2], [6,6], [8,7], [7,5],[7,6], [7,7], [9,0], [9,3]
+],
+3 : [
+    [0,0], [0,1], [0,7], [0,6], [1,0], [1,3], [2,2], [3,7], [3,9], [4,2], [4,5], [4,6], [4,9], [5,0], [5,1], [5,2], [5,9], [6,5], [8,0], [8,8], [9,0], [9,1], [9,4], [9,5]
+],
+4 : [
+    [0,3], [0,7], [0,8], [0,9], [1,1], [1,9], [2,1], [2,5], [2,7], [2,9], [3,5], [3,9], [4,0], [4,4], [4,5], [5,9], [7,0], [8,6], [9,4],[9,5], [9,6], [9,9]
+],
+0 : [
+    [0,0], [0,1], [0,2], [0,3], [0,4], [0,6], [0,7], [0,8], [0,9], [1,0], [1,6], [1,7], [1,9], [2,0], [2,2], [2,3], [2,4], [2,5], [2,6], [2,7], [2,9], [3,0], [3,4], [4,0], [4,1], [4,2], [4,4], [4,6], [4,7], [5,0], [5,1], [5,2], [5,6], [5,7], [6,0], [6,1], [6,2], [6,3], [6,4], [6,5], [7,0], [7,7], [7,9], [8,0], [8,9], [9,0], [9,1], [9,2], [9,3], [9,4], [9,5], [9,6], [9,7], [9,8], [9,9]
+]
+}
+playerpos = {
+1 : [3,5],
+2 : [3,2],
+3 : [9,9],
+4 : [4,9],
+0 : [1,8]
+}
+minmoves = {
+1 : 43,
+2 : 45,
+3 : 40,
+4 : 32,
+0 : 15
+}
