+++ walls.py
@@ -1,0 +1,6 @@
+class walls(object):
+    
+    def __init__(self):
+        
+        self.locations = []
+        self.count = 0
